refactor: create MainPage component
