--- src/App.css
@@ -1,4 +1,0 @@
-.throw-error-button {
-  background-color: rgb(246, 175, 175);
-  margin-bottom: 10px;
-}
