+++ src/components/pages/main_page/MainPage.css
@@ -1,0 +1,4 @@
+.throw-error-button {
+    background-color: rgb(246, 175, 175);
+    margin-bottom: 10px;
+  }
